refactor: move material synthesis flow to processing task
